Changed PETSc options to choose AMD for the permutation when factorizing with MUMPS. The default option is Metis and this produces a segfault when used in the expansion driver. I think this is because in this driver we are also invoking Metis through Metis.jl as a dependency of GridapGmsh to partition the model.
There are therefore two verstions of Metis with, I think, different integer presicion.
This happens using the compiled version of PETSc (with JULIA_PETSC_LIBRARY non empty)
which uses Int64. I have not found how to  control the integer presicion in Metis.jl,
it seems to be automatically chosen by METIS_jll.jl and I guess it is Int32.

I have also noticed that the default GridapPETSc installation (with JULIA_PETSC_LIBRARY empty)
uses Int32 by default. I will try this instalation to check that this is the origin of the
segfault.